+++ aldryn_forms/utils.py
@@ -1,15 +1,36 @@
+import logging
+import smtplib
+from typing import TYPE_CHECKING, NamedTuple
+
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.forms.forms import NON_FIELD_ERRORS
 from django.utils.module_loading import import_string
+from django.utils.translation import get_language
 
 from cms.cms_plugins import AliasPlugin
 from cms.utils.moderator import get_cmsplugin_queryset
 from cms.utils.plugins import downcast_plugins
 
+from emailit.api import send_mail
+from emailit.utils import get_template_names
+
 from .action_backends_base import BaseAction
 from .compat import build_plugin_tree
 from .constants import ALDRYN_FORMS_ACTION_BACKEND_KEY_MAX_SIZE, DEFAULT_ALDRYN_FORMS_ACTION_BACKENDS
+from .validators import is_valid_recipient
+
+
+if TYPE_CHECKING:
+    from .models import FormSubmissionBase
+
+
+class NameTypeField(NamedTuple):
+    name: str
+    value: str
+
+
+logger = logging.getLogger(__name__)
 
 
 def get_action_backends():
@@ -120,3 +141,35 @@ def add_form_error(form, message, field=NON_FIELD_ERRORS):
         form._errors[field].append(message)
     except KeyError:
         form._errors[field] = form.error_class([message])
+
+
+def send_postponed_notifications(instance: "FormSubmissionBase") -> bool:
+    """Send postponed notifications."""
+    recipients = [user for user in instance.get_recipients() if is_valid_recipient(user.email)]
+    form_data = [NameTypeField(item.name, item.value) for item in instance.get_form_data()]
+    context = {
+        'form_name': instance.name,
+        'form_data': form_data,
+        'form_plugin': instance,
+    }
+    subject_template_base = getattr(settings, 'ALDRYN_FORMS_EMAIL_SUBJECT_TEMPLATES_BASE',
+                                    getattr(settings, 'ALDRYN_FORMS_EMAIL_TEMPLATES_BASE', None))
+    if subject_template_base:
+        language = instance.language or get_language()
+        subject_templates = get_template_names(language, subject_template_base, 'subject', 'txt')
+    else:
+        subject_templates = None
+
+    try:
+        send_mail(
+            recipients=[user.email for user in recipients],
+            context=context,
+            template_base=getattr(
+                settings, 'ALDRYN_FORMS_EMAIL_TEMPLATES_BASE', 'aldryn_forms/emails/notification'),
+            subject_templates=subject_templates,
+            language=instance.language,
+        )
+    except smtplib.SMTPException as err:
+        logger.error(err)
+        return False
+    return True